requirements.txt
python-dotenv
pandas
numpy
matplotlib
seaborn
flask
pymango
certifi
scikit-learn
pymongo[srv]==3.12

# -e .